[why] fix compute [who] huangyizhi

+++ main.cpp
@@ -9,11 +9,11 @@
 
 std::unordered_map<int, XPU_INFO> MF::MFServer::worker_xpu_info;
 MF::DataManager MF::MFServer::dm("netflix_train.bin", 128, 3);
-size_t MF::MFServer::cpus(0);
-size_t MF::MFServer::gpus(0);
-size_t MF::MFServer::fpgas(0);
-size_t MF::MFServer::tpus(0);
-size_t MF::MFServer::xpus(0);
+int MF::MFServer::cpus(0);
+int MF::MFServer::gpus(0);
+int MF::MFServer::fpgas(0);
+int MF::MFServer::tpus(0);
+int MF::MFServer::xpus(0);
 int MF::MFServer::max_workers(0);
 int MF::MFServer::scale(0);
 int MF::MFServer::nr_threads(0);
